T127: Major refactor.
