+++ apps/app_test_symmetry.cpp
@@ -4,7 +4,7 @@
  * By Sebastian Raaphorst, 2018.
  */
 
-#include <types/Symmetry.h>
+#include <types/Transformation.h>
 #include <typeclasses/Show.h>
 #include <maze/DFSMazeGenerator.h>
 #include <maze/Maze.h>
@@ -15,7 +15,7 @@ int main(int argc, char *argv[]) {
     maze::DFSMazeGenerator dfs(50, 40);
     maze::Maze m = dfs.generate();
     std::cout << typeclasses::Show<maze::Maze>::show(m);
-    maze::Maze m90 = m.applySymmetry(types::Symmetry::IDENTITY);
+    maze::Maze m90 = m.applyTransformation(types::Transformation::IDENTITY);
     std::cout << typeclasses::Show<maze::Maze>::show(m90);
     return 0;
 }
