+++ gui/MazeWidget.h
@@ -27,9 +27,10 @@ namespace spelunker::gui {
 
         maze::Maze maze;
 
-        static auto const FLOOR_COLOUR     = Qt::gray;
+        static auto const FLOOR_COLOUR1    = Qt::gray;
+        static auto const FLOOR_COLOUR2    = Qt::lightGray;
         static auto const WALL_COLOUR      = Qt::black;
         static auto constexpr WALL_WIDTH   = 1;
-        static auto constexpr BORDER_WIDTH = 5;
+        static auto constexpr BORDER_WIDTH = 3;
     };
 }
